+++ node/resolvers/orderForm.ts
@@ -1,8 +1,9 @@
+import { MutationUpdateOrderFormProfileArgs } from 'vtex.checkout-graphql'
+
 import { Clients } from '../clients'
 import { adjustItems } from './items'
 import { fillMessages } from './messages'
 import { getShippingInfo } from './shipping/utils/shipping'
-import { DEFAULT_ORDER_FORM_SECTIONS } from '../constants'
 
 export const getNewOrderForm = async ({
   clients,
@@ -62,7 +63,7 @@ export const queries = {
 export const mutations = {
   updateOrderFormProfile: async (
     _: unknown,
-    { email }: { email: string },
+    { input }: MutationUpdateOrderFormProfileArgs,
     ctx: Context
   ): Promise<OrderForm> => {
     const {
@@ -72,7 +73,7 @@ export const mutations = {
 
     const orderFormWithProfile = await checkout.updateOrderFormProfile(
       orderFormId!,
-      { email, expectedOrderFormSections: DEFAULT_ORDER_FORM_SECTIONS }
+      input
     )
 
     return orderFormWithProfile