+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Chore
+
+- Update `updateOrderFormProfile` mutation input to receive all user profile fields, not just email.
+
 ## [0.24.0] - 2020-03-02
 
 ### Added